feat(abell): add serverBuild option support to configure only the server build
