+++ packages/abell/src/type-utils.ts
@@ -1,3 +1,8 @@
+import type {
+  EsbuildTransformOptions,
+  UserConfig as ViteUserConfig
+} from 'vite';
+
 type RenderContext = Record<string, unknown>;
 
 export type Route = {
@@ -23,3 +28,31 @@ export type AbstractSyntaxArrayType = {
     blocks: { text: string }[];
   };
 };
+
+export type AbellOptions = {
+  /**
+   * This is configurations for the esbuild that transforms internal abell blocks.
+   *
+   * E.g. if you want to use JSX in abell blocks, you can override loader of esbuild here
+   * ```js
+   * abell: {
+   *  esbuild: {
+   *    loader: 'jsx'
+   *  }
+   * }
+   * ```
+   */
+  esbuild?: EsbuildTransformOptions;
+  /**
+   * Just like "build" option in Vite. Except this only applies on intermediate server build that is created.
+   *
+   * You can use this to change config of build-time things.
+   *
+   * E.g. if you want to use top-level await in entry.build.ts, you can set target here without having to change target of your final client bundle
+   */
+  serverBuild?: ViteUserConfig['build'];
+};
+
+export interface AbellViteConfig extends ViteUserConfig {
+  abell?: AbellOptions;
+}
